refactor(env): move Env model to env_model.py to prevent circular import
