+++ analyst_agent/react/graph.py
@@ -9,7 +9,7 @@ from langgraph.checkpoint.memory import MemorySaver
 from queue import Queue
 from langchain_core.runnables import RunnableConfig  
 from analyst_agent.react.base import BaseNode
-from analyst_agent.react.logger import RunLogger
+from analyst_agent.react.env_model import Env
 
 
 #TODO bring csv file path and create methods to read csv file in codeexecutornode.
@@ -17,13 +17,13 @@ class ChartAgentExecutorNode(BaseNode):
     '''
     차트 생성 agent를 실행시키는 node.
     '''
-    def __init__(self, verbose=False, queue: Queue=None, run_logger: RunLogger=None, **kwargs):
-        super().__init__(verbose=verbose, queue=queue, run_logger=run_logger, **kwargs)
+    def __init__(self, verbose=False, env: Env=None, queue: Queue=None, **kwargs):
+        super().__init__(verbose=verbose, env=env, queue=queue, **kwargs)
 
     
     def run(self, state: AgentContextState):
         config = RunnableConfig(recursion_limit=5) 
-        chart_graph = chart_code_react_agent(queue=self.queue, run_logger=self.run_logger)
+        chart_graph = chart_code_react_agent(queue=self.queue, env=self.env)
         result : ChartState = chart_graph.invoke(
             input={
                 'user_query' : state['user_query'], 
@@ -39,14 +39,14 @@ class DataFrameAgentExecutorNode(BaseNode):
     '''
     차트 생성 agent를 실행시키는 node.
     '''
-    def __init__(self, verbose=False, queue: Queue=None, run_logger: RunLogger=None, **kwargs):
-        super().__init__(verbose=verbose, queue=queue, run_logger=run_logger, **kwargs)
+    def __init__(self, verbose=False, env: Env=None, queue: Queue=None, **kwargs):
+        super().__init__(verbose=verbose, env=env, queue=queue, **kwargs)
 
     #TODO bring csv file path and create methods to read csv file in codeexecutornode.
     def run(self, state: AgentContextState):
         
         config = RunnableConfig(recursion_limit=5) 
-        df_graph = df_code_react_agent(queue=self.queue, run_logger=self.run_logger)
+        df_graph = df_code_react_agent(queue=self.queue, env=self.env)
         result : DataFrameState = df_graph.invoke(
             input={
                 'user_query' : state['user_query'], 
@@ -65,9 +65,9 @@ class DataFrameAgentExecutorNode(BaseNode):
         state['dataframe_code'] = result['df_code']
         return state
 
-def chart_code_react_agent(verbose: bool = False, track_time: bool = False, queue: Queue=None, run_logger: RunLogger=None) -> CompiledStateGraph:
-    chart_code_generator_node = ChartCodeGeneratorNode(verbose=verbose, track_time=track_time, queue=queue, run_logger=run_logger)
-    chart_code_executor_node = ChartCodeExecutorNode(verbose=verbose, track_time=track_time, queue=queue, run_logger=run_logger)
+def chart_code_react_agent(verbose: bool = False, track_time: bool = False, queue: Queue=None, env: Env=None) -> CompiledStateGraph:
+    chart_code_generator_node = ChartCodeGeneratorNode(verbose=verbose, track_time=track_time, queue=queue, env=env)
+    chart_code_executor_node = ChartCodeExecutorNode(verbose=verbose, track_time=track_time, queue=queue, env=env)
     
     chart_code_agent_workflow = StateGraph(ChartState)
     chart_code_agent_workflow.add_node('chart_code_generator', chart_code_generator_node)
@@ -78,9 +78,9 @@ def chart_code_react_agent(verbose: bool = False, track_time: bool = False, queu
     chart_memory = MemorySaver()
     return chart_code_agent_workflow.compile(checkpointer=chart_memory)
 
-def df_code_react_agent(verbose: bool = False, track_time: bool = False, queue: Queue=None, run_logger: RunLogger=None) -> CompiledStateGraph:
-    dataframe_code_generator_node = DataFrameCodeGeneratorNode(verbose=verbose, track_time=track_time, queue=queue, run_logger=run_logger)
-    dataframe_code_executor_node = DataFrameCodeExecutorNode(verbose=verbose, track_time=track_time, queue=queue, run_logger=run_logger)
+def df_code_react_agent(verbose: bool = False, track_time: bool = False, queue: Queue=None, env: Env=None) -> CompiledStateGraph:
+    dataframe_code_generator_node = DataFrameCodeGeneratorNode(verbose=verbose, track_time=track_time, queue=queue, env=env)
+    dataframe_code_executor_node = DataFrameCodeExecutorNode(verbose=verbose, track_time=track_time, queue=queue, env=env)
     
     dataframe_code_agent_workflow = StateGraph(DataFrameState)
     dataframe_code_agent_workflow.add_node('dataframe_code_generator', dataframe_code_generator_node)
@@ -92,10 +92,10 @@ def df_code_react_agent(verbose: bool = False, track_time: bool = False, queue: 
     return dataframe_code_agent_workflow.compile(checkpointer=dataframe_memory)
 
 
-def react_code_agent(verbose: bool = False, track_time: bool = False, queue: Queue=None, run_logger: RunLogger=None) -> CompiledStateGraph:
-    router_node = RouterNode(verbose=verbose, track_time=track_time, queue=queue, run_logger=run_logger)
-    dataframe_code_agent = DataFrameAgentExecutorNode(verbose=verbose, track_time=track_time, queue=queue, run_logger=run_logger)
-    chart_code_agent = ChartAgentExecutorNode(verbose=verbose, track_time=track_time, queue=queue, run_logger=run_logger)
+def react_code_agent(verbose: bool = False, track_time: bool = False, queue: Queue=None, env: Env=None) -> CompiledStateGraph:
+    router_node = RouterNode(verbose=verbose, track_time=track_time, queue=queue, env=env)
+    dataframe_code_agent = DataFrameAgentExecutorNode(verbose=verbose, track_time=track_time, queue=queue, env=env)
+    chart_code_agent = ChartAgentExecutorNode(verbose=verbose, track_time=track_time, queue=queue, env=env)
     
     react_code_agent_workflow = StateGraph(AgentContextState)
     react_code_agent_workflow.add_node('router', router_node)
